Module Condition_map.ConditionMapTree

One implementation of LConditionMap, where partitions are hierarchically split in two, based on the order in which they were inserted.

Parameters

module Cond : CONDITION

Signature

type 'a t

module Make (L : L) : LConditionMap with module L = L and module Cond = Cond and type t = L.t t